Refactor logging to be a package, instead of a module
